Nick's changes, 10/19
Copied Level1 Scene into Nick's folder for testing purposes.

Added and applied tags for Enemy and Collectible.

Created and applied script to Enemy prefab that destroys only the player upon contact. This does not reflect how it will play in the final version if the game, but will suffice as a placeholder.